+++ agent/src/collector/collector.rs
@@ -766,11 +766,14 @@ impl Stash {
 
         // 双端统计量不要对Meter反向
         let key = StashKey::new(&tagger, src_ip, Some(dst_ip));
-        self.add(
-            key,
-            tagger.clone(),
-            Meter::Flow(acc_flow.flow_meter.clone()),
-        );
+        if !flow_key.tap_port.is_from_ebpf() {
+            // the data from ebpf has not l4 info, it doesn't have flow_meter
+            self.add(
+                key,
+                tagger.clone(),
+                Meter::Flow(acc_flow.flow_meter.clone()),
+            );
+        }
         if tagger.l7_protocol != L7Protocol::Unknown
             && self.context.config.load().l7_metrics_enabled
         {
