+++ agent/src/collector/quadruple_generator.rs
@@ -366,6 +366,10 @@ impl SubQuadGen {
         possible_host: &mut PossibleHost,
     ) {
         for acc_flow in flows.iter_mut() {
+            if acc_flow.tagged_flow.flow.flow_key.tap_port.is_from_ebpf() {
+                //  the data from ebpf has not l4 info, it needn't to set_connection
+                continue;
+            }
             acc_flow.is_active_host0 = Self::check_active_host(
                 possible_host,
                 &acc_flow.tagged_flow.flow.flow_metrics_peers[0],
@@ -542,7 +546,7 @@ pub struct QuadrupleGeneratorThread {
     input: Arc<Receiver<Box<TaggedFlow>>>,
     second_output: DebugSender<Box<AccumulatedFlow>>,
     minute_output: DebugSender<Box<AccumulatedFlow>>,
-    flow_output: DebugSender<Arc<TaggedFlow>>,
+    flow_output: Option<DebugSender<Arc<TaggedFlow>>>, // flow_output sends TaggedFlows to the FlowAggr, it will be None when the QuadrupleGeneratorThread is created by the ebpf_dispatcher
     connection_lru_capacity: usize,
     metrics_type: MetricsType,
     second_delay_seconds: u64,
@@ -567,7 +571,7 @@ impl QuadrupleGeneratorThread {
         input: Receiver<Box<TaggedFlow>>,
         second_output: DebugSender<Box<AccumulatedFlow>>,
         minute_output: DebugSender<Box<AccumulatedFlow>>,
-        flow_output: DebugSender<Arc<TaggedFlow>>,
+        flow_output: Option<DebugSender<Arc<TaggedFlow>>>,
         connection_lru_capacity: usize,
         metrics_type: MetricsType,
         second_delay_seconds: u64,
@@ -583,7 +587,7 @@ impl QuadrupleGeneratorThread {
             input: Arc::new(input),
             second_output: second_output.clone(),
             minute_output: minute_output.clone(),
-            flow_output: flow_output.clone(),
+            flow_output,
             connection_lru_capacity,
             metrics_type,
             second_delay_seconds,
@@ -686,7 +690,7 @@ pub struct QuadrupleGenerator {
 
     key: QgKey,
     policy_ids: [U16Set; 2],
-    output_flow: DebugSender<Arc<TaggedFlow>>,
+    output_flow: Option<DebugSender<Arc<TaggedFlow>>>, // it will be None when the QuadrupleGeneratorThread is created by the ebpf_dispatcher
 
     l7_metrics_enabled: Arc<AtomicBool>,
     vtap_flow_1s_enabled: Arc<AtomicBool>,
@@ -704,7 +708,7 @@ impl QuadrupleGenerator {
         input: Arc<Receiver<Box<TaggedFlow>>>,
         second_output: DebugSender<Box<AccumulatedFlow>>,
         minute_output: DebugSender<Box<AccumulatedFlow>>,
-        flow_output: DebugSender<Arc<TaggedFlow>>,
+        flow_output: Option<DebugSender<Arc<TaggedFlow>>>,
         connection_lru_capacity: usize,
         metrics_type: MetricsType,
         second_delay_seconds: u64,
@@ -1107,8 +1111,11 @@ impl QuadrupleGenerator {
             match self.input.recv(Some(Duration::from_secs(3))) {
                 Ok(tagged_flow) => {
                     let tagged_flow = Arc::new(*tagged_flow);
-                    if let Err(_) = self.output_flow.send(tagged_flow.clone()) {
-                        debug!("qg push tagged flows to l4_flow queue failed maybe queue have terminated");
+                    if self.output_flow.is_some() {
+                        if let Err(_) = self.output_flow.as_mut().unwrap().send(tagged_flow.clone())
+                        {
+                            debug!("qg push tagged flows to l4_flow queue failed maybe queue have terminated");
+                        }
                     }
                     if self.collector_enabled.load(Ordering::Relaxed) {
                         self.handle(Some(tagged_flow.clone()), tagged_flow.flow.flow_stat_time);
