+++ agent/src/dispatcher/analyzer_mode_dispatcher.rs
@@ -119,8 +119,9 @@ impl AnalyzerModeDispatcher {
             base.ntp_diff.clone(),
             base.flow_map_config.clone(),
             base.log_parse_config.clone(),
-            base.packet_sequence_output_queue.clone(), // Enterprise Edition Feature: packet-sequence
+            Some(base.packet_sequence_output_queue.clone()), // Enterprise Edition Feature: packet-sequence
             &base.stats,
+            false, // $from_ebpf: false, packets aren't from ebpf, it needs to collect packet sequence data
         );
 
         while !base.terminated.load(Ordering::Relaxed) {
