+++ agent/src/dispatcher/local_mode_dispatcher.rs
@@ -70,8 +70,9 @@ impl LocalModeDispatcher {
             base.ntp_diff.clone(),
             base.flow_map_config.clone(),
             base.log_parse_config.clone(),
-            base.packet_sequence_output_queue.clone(), // Enterprise Edition Feature: packet-sequence
+            Some(base.packet_sequence_output_queue.clone()), // Enterprise Edition Feature: packet-sequence
             &base.stats,
+            false, // $from_ebpf: false, packets aren't from ebpf, it needs to collect packet sequence data
         );
 
         while !base.terminated.load(Ordering::Relaxed) {
